convert burning date to milliseconds

+++ bonusesdata/src/main/java/com/grizzlyorange/bonusesdata/api/data/BonusesInfo.kt
@@ -3,5 +3,5 @@ package com.grizzlyorange.bonusesdata.api.data
 data class BonusesInfo(
     val totalBonusesAmount: Int = 0,
     val burningBonusesAmount: Int = 0,
-    val burningDateInMillis: Int = 0
+    val burningDateInMillis: Long = 0
 )
